Deadlock prevents "./imapsync list" from working
The 'connect' function expected there to be someone listening on the
end of the statusChan, which for the "list" operation wasn't actually
the case. Factored the common listener code out into "reportOnStatus"
function.

Now the whole thing works -- both sees my Gmail labels and is able to
download to .mbox files.